+++ packages/lit/tsup.config.ts
@@ -3,6 +3,7 @@ export default defineConfig({
   dts: true,
   entry: {
     lit: 'src/index.ts',
+    'lit.transform': 'src/transform.ts',
     'lit.adapter.server': 'src/adapter.server.ts',
     'lit.adapter.client': 'src/adapter.client.ts',
   },